+++ privacyidea/static_new/src/app/components/external-services/privacyidea-servers/privacyidea-servers.component.ts
@@ -24,7 +24,11 @@ import { MatSort, MatSortModule } from "@angular/material/sort";
 import { MatIconModule } from "@angular/material/icon";
 import { MatButtonModule } from "@angular/material/button";
 import { MatDialog, MatDialogModule } from "@angular/material/dialog";
-import { PrivacyideaServer, PrivacyideaServerService, PrivacyideaServerServiceInterface } from "../../../services/privacyidea-server/privacyidea-server.service";
+import {
+  PrivacyideaServer,
+  PrivacyideaServerService,
+  PrivacyideaServerServiceInterface
+} from "../../../services/privacyidea-server/privacyidea-server.service";
 import { NewPrivacyideaServerComponent } from "./new-privacyidea-server/new-privacyidea-server.component";
 import { AuthService, AuthServiceInterface } from "../../../services/auth/auth.service";
 import { MatTooltipModule } from "@angular/material/tooltip";
@@ -35,6 +39,7 @@ import { MatFormField, MatInput, MatLabel } from "@angular/material/input";
 import { ClearableInputComponent } from "../../shared/clearable-input/clearable-input.component";
 import { TableUtilsService, TableUtilsServiceInterface } from "../../../services/table-utils/table-utils.service";
 import { CopyButtonComponent } from "../../shared/copy-button/copy-button.component";
+import { SimpleConfirmationDialogComponent } from "@components/shared/dialog/confirmation-dialog/confirmation-dialog.component";
 
 @Component({
   selector: "app-privacyidea-servers",
@@ -93,18 +98,22 @@ export class PrivacyideaServersComponent {
   }
 
   deleteServer(server: PrivacyideaServer): void {
-    this.dialogService.confirm({
-      data: {
-        title: $localize`Delete privacyIDEA Server`,
-        serialList: [server.identifier],
-        type: "privacyidea-server",
-        action: "delete"
-      }
-    }).then(result => {
-      if (result) {
-        this.privacyideaServerService.deletePrivacyideaServer(server.identifier);
-      }
-    });
+    this.dialogService
+      .openDialog({
+        component: SimpleConfirmationDialogComponent,
+        data: {
+          title: $localize`Delete privacyIDEA Server`,
+          items: [server.identifier],
+          itemType: "privacyidea-server",
+          confirmAction: { label: $localize`Delete`, value: true, type: "destruct" }
+        }
+      })
+      .afterClosed()
+      .subscribe({
+        next: (result) => {
+          if (result) this.privacyideaServerService.deletePrivacyideaServer(server.identifier);
+        }
+      });
   }
 
   onFilterInput(value: string): void {
